requirements.txt
flask==2.3.3
requests==2.31.0
openai==1.1.1
langchain==0.0.292
langchain-openai==0.0.2.post1
langchain-core==0.1.1
langchain-community==0.0.10
python-dotenv==1.0.0
twilio==8.5.0
pydantic==2.4.2
python-decouple==3.8
pytz==2023.3
loguru==0.7.2
gunicorn==21.2.0
pyngrok==6.1.0
numpy==1.24.3
scikit-learn==1.3.0
pydantic-settings==2.0.3